Split debugger from main emu run loop
